Add Insert Many Operation (#4)
This change adds an `insert_many` helper function in addition to the
existing `insert` operation. SQLAlchemy implicitly allows insertmany
when using the `sqlalchemy.Insert` statement, but the api for the `insert`
helper in grammdb does not allow passing a list of values, so this
functionality was not accessible in the grammdb wrapper functions.

I chose to add an `insert_many` function with a slightly different
interface from the `insert` function to make it explicit when a query is
intended to do a bulk insert.